+++ src/app/pages/demo-components/filter-components/filter-components.component.ts
@@ -19,7 +19,7 @@ export class FilterComponentsComponent implements OnInit {
       gold: '#b6ab66'
   };
 
-components = `<!-- Filter Checkboxes Component - Multiple Group -->
+checkboxComponent = `<!-- Filter Checkboxes Component - Multiple Group -->
 <mds-filter-checkbox
   [titlegroup]="'Filter by'"
   [filterData]="sampledata"
@@ -27,9 +27,9 @@ components = `<!-- Filter Checkboxes Component - Multiple Group -->
   [filterSelected]="selected"
   [reset]="'✕'"
   [hideCounter]="false"
-></mds-filter-checkbox>
+></mds-filter-checkbox>`;
 
-<!-- Filter Swatch Component -->
+swatchComponent = `<!-- Filter Swatch Component -->
 <mds-filter-swatch
   [titlegroup]="'Filter by Color'"
   [filterData]="sampledata"
@@ -38,18 +38,12 @@ components = `<!-- Filter Checkboxes Component - Multiple Group -->
   [swatchMapping]="colormap"
   [swatchSize]="30"
   [swatchRadius]="25"
-></mds-filter-swatch>
+></mds-filter-swatch>`;
 
-<!-- Filter Checkboxes Component - Single Group -->
-<mds-filter-checkbox
-  [filterData]="sampledata"
-  [filterBy]="['brand']"
-  [filterSelected]="selected"
-  [hideCounter]="false"
-></mds-filter-checkbox>
-
-<!-- Reset Filter Component -->
+resetComponent = `<!-- Reset Filter Component -->
 <mds-filter-reset [content]="'Reset All'" [(filterSelected)]="selected"></mds-filter-reset>`;
+
+
 importmodule = `import { MdsFilterModule } from 'medes-ui';
 
 @NgModule({
@@ -57,7 +51,9 @@ importmodule = `import { MdsFilterModule } from 'medes-ui';
   imports: [
     MdsFilterModule
   ]
-})`;
+})
+
+export class MyModule { }`;
 
   ngOnInit(): void {
     this.sampledata = SampleProductsData.data;