+++ src/app/pages/demo-ui-boilerplate/ui-boilerplate-routing.module.ts
@@ -3,12 +3,14 @@ import { RouterModule, Routes } from '@angular/router';
 import { GridSystemComponent } from './grid-system/grid-system.component';
 import { ImageLoaderComponent } from './image-loader/image-loader.component';
 import { MdsDocBoilerplateComponent } from './ui-boilerplate.component';
+import { LayoutUtilsComponent } from './layout-utils/layout-utils.component';
 
 const routes: Routes = [
   { path: '', redirectTo: 'basic-elements' },
   { path: 'basic-elements', component: MdsDocBoilerplateComponent },
-  { path: 'image-loader', component: ImageLoaderComponent },
-  { path: 'grid-system', component: GridSystemComponent }
+  { path: 'grid-system', component: GridSystemComponent },
+  { path: 'layout-utils', component: LayoutUtilsComponent },
+  { path: 'image-loader', component: ImageLoaderComponent }
 ];
 
 @NgModule({