+++ src/app/pages/demo-utils/number-utils/number-utils.component.ts
@@ -1,8 +1,5 @@
 import { Component } from '@angular/core';
 
-// import { MdsNumberUtils } from 'medes-ui';
-// import { MdsNumberUtils} from 'projects/medes-ui/src/public-api';
-
 @Component({
   selector: 'mds-number-utils',
   templateUrl: './number-utils.component.html',